+++ src/main/java/com/wildfire/main/WildfireGender.java
@@ -20,38 +20,48 @@ package com.wildfire.main;
 
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Optional;
 import java.util.UUID;
-import java.util.concurrent.Future;
-
 import com.mojang.logging.LogUtils;
 import com.wildfire.main.entitydata.PlayerConfig;
-import net.fabricmc.api.ClientModInitializer;
-import net.minecraft.util.Util;
+import com.wildfire.main.networking.WildfireSync;
+import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.networking.v1.EntityTrackingEvents;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.server.network.ServerPlayerEntity;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 
-public class WildfireGender implements ClientModInitializer {
-	public static final String VERSION = "3.1";
-  	public static final String MODID = "wildfire_gender";
+public class WildfireGender implements ModInitializer {
+	public static final String MODID = "wildfire_gender";
 	public static final Logger LOGGER = LogUtils.getLogger();
 	public static final Map<UUID, PlayerConfig> PLAYER_CACHE = new HashMap<>();
 
 	@Override
-  	public void onInitializeClient() {
-		WildfireEventHandler.registerClientEvents();
-    }
+	public void onInitialize() {
+		WildfireSync.register();
+		EntityTrackingEvents.START_TRACKING.register(this::onBeginTracking);
+	}
 
 	public static @Nullable PlayerConfig getPlayerById(UUID id) {
-		  return PLAYER_CACHE.get(id);
+		return PLAYER_CACHE.get(id);
 	}
 
 	public static @NotNull PlayerConfig getOrAddPlayerById(UUID id) {
 		return PLAYER_CACHE.computeIfAbsent(id, PlayerConfig::new);
 	}
 
-  	public static Future<Optional<PlayerConfig>> loadGenderInfo(UUID uuid, boolean markForSync) {
-	    return Util.getIoWorkerExecutor().submit(() -> Optional.ofNullable(PlayerConfig.loadCachedPlayer(uuid, markForSync)));
-  	}
+	private void onBeginTracking(Entity tracked, ServerPlayerEntity syncTo) {
+		if(tracked instanceof PlayerEntity toSync) {
+			PlayerConfig genderToSync = WildfireGender.getPlayerById(toSync.getUuid());
+			if(genderToSync == null) return;
+			// Note that we intentionally don't check if we've previously synced a player with this code path;
+			// because we use entity tracking to sync, it's entirely possible that one player would leave the
+			// tracking distance of another, change their settings, and then re-enter their tracking distance;
+			// we wouldn't sync while they're out of tracking distance, and as such, their settings would be out
+			// of sync until they relog.
+			WildfireSync.sendToClient(syncTo, genderToSync);
+		}
+	}
 }
