+++ src/main/resources/wildfire_gender.mixins.json
@@ -4,13 +4,13 @@
   "package": "com.wildfire.mixins",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
-    "ArmorStandEntityMixin",
-    "LivingEntityMixin"
+    "ArmorStandEntityMixin"
   ],
   "client": [
     "ArmorStandEntityRendererMixin",
     "PlayerRenderMixin",
-    "BreastPhysicsTickMixin"
+    "BreastPhysicsTickMixin",
+    "LivingEntityMixin"
   ],
   "injectors": {
     "defaultRequire": 1
